App works but have a couple of test and bugs to handle.

+++ app/src/main/java/com/example/mymusicplayer/data/source/remote/model/LearnFieldSongs.kt
@@ -21,7 +21,8 @@ fun LearnFieldSongsItems.toSong(): Song {
         Song(
             title = title,
             coverUrl = cover,
-            totalDurationMs = totalDurationMs,
+            totalDurationMs =
+            totalDurationMs.toLong(),
             audioUrl = audio,
             isFavourite = false,
             pathToAudio = "",
